telegram-reminderbot/app/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-telegram-bot>=13
munch==2.5.0
pymongo[srv]==3.11
APScheduler==3.6.3
requests==2.26.0
python-telegram-bot-calendar==1.0.5